+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- CS-460 - Added frontend scout csv import-export
+- CS-460 - Fixed getBirthDateFromCNP method
+- CS-460 - Fixed typo in getOptionsWithLabels method name
 - CS-462 - Added warning to team report if any scout doesn't have registration form uploaded
 - CS-436 - Added option to hide mandate types on frontend
 - CS-466 - Created migrations for "Knowledge repository" models